test: split runtime rewind tests
