+++ tests/test_runtime_orchestrator.py
@@ -13,7 +13,6 @@ from forgeflow_runtime.orchestrator import (
     retry_stage,
     run_route,
     status_summary,
-    step_back,
 )
 from forgeflow_runtime.generator import GenerationError
 
@@ -575,207 +574,6 @@ def test_large_route_rejects_missing_eval_record_before_long_run(tmp_path: Path)
     with pytest.raises(RuntimeViolation, match="long-run requires artifacts satisfying gate worth_long_run_capture: eval-record"):
         run_route(task_dir=task_dir, policy=policy, route_name="large_high_risk")
 
-
-def test_step_back_rewinds_to_previous_stage(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    state = step_back(task_dir=task_dir, policy=policy, route_name="small", current_stage="quality-review")
-
-    assert state["current_stage"] == "execute"
-    assert state["status"] == "in_progress"
-    checkpoint = json.loads((task_dir / "checkpoint.json").read_text(encoding="utf-8"))
-    _assert_schema_valid("checkpoint", checkpoint)
-    assert checkpoint["route"] == "small"
-    assert checkpoint["current_stage"] == "execute"
-    assert checkpoint["next_action"] == "Resume at quality-review after reloading canonical artifacts."
-
-
-def test_step_back_rewinds_plan_ledger_progress_for_medium_route(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_medium_plan_artifacts(task_dir)
-    _write_json(
-        task_dir / "run-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "current_stage": "quality-review",
-            "status": "in_progress",
-            "completed_gates": ["clarification_complete"],
-            "failed_gates": [],
-            "retries": {},
-            "current_task_id": "task-1",
-            "spec_review_approved": False,
-            "quality_review_approved": False,
-        },
-    )
-    _write_json(
-        task_dir / "plan-ledger.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "medium",
-            "completed_stages": ["clarify", "plan", "execute"],
-            "completed_gates": ["clarification_complete", "plan_executable", "execution_evidenced"],
-            "retries": {},
-            "current_task_id": "task-1",
-            "tasks": [
-                {
-                    "id": "task-1",
-                    "title": "update workflow docs",
-                    "depends_on": [],
-                    "files": ["docs/workflow.md"],
-                    "parallel_safe": False,
-                    "status": "in_progress",
-                    "required_gates": ["machine", "validator"],
-                    "evidence_refs": [
-                        "run-state.json#gate:plan_executable",
-                        "run-state.json#gate:execution_evidenced",
-                    ],
-                    "attempt_count": 0,
-                }
-            ],
-        },
-    )
-    _write_json(
-        task_dir / "checkpoint.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "medium",
-            "current_stage": "quality-review",
-            "current_task_id": "task-1",
-            "plan_ref": "plan.json",
-            "plan_ledger_ref": "plan-ledger.json",
-            "run_state_ref": "run-state.json",
-            "latest_review_ref": "review-report.json",
-            "next_action": "Resume at finalize after review approval.",
-            "open_blockers": [],
-            "updated_at": "2026-04-22T00:00:00Z",
-        },
-    )
-    _write_json(
-        task_dir / "session-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "medium",
-            "current_stage": "quality-review",
-            "current_task_id": "task-1",
-            "plan_ref": "plan.json",
-            "plan_ledger_ref": "plan-ledger.json",
-            "run_state_ref": "run-state.json",
-            "latest_checkpoint_ref": "checkpoint.json",
-            "latest_review_ref": "review-report.json",
-            "next_action": "Resume at finalize after review approval.",
-            "updated_at": "2026-04-22T00:00:00Z",
-        },
-    )
-    _write_json(
-        task_dir / "review-report.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "review_type": "quality",
-            "verdict": "approved",
-            "findings": ["looks fine"],
-            "approved_by": "quality-reviewer",
-            "next_action": "finalize 가능",
-        },
-    )
-
-    state = step_back(task_dir=task_dir, policy=policy, route_name="medium", current_stage="quality-review")
-
-    assert state["current_stage"] == "execute"
-    persisted_plan_ledger = json.loads((task_dir / "plan-ledger.json").read_text(encoding="utf-8"))
-    assert persisted_plan_ledger["completed_stages"] == ["clarify", "plan"]
-    assert persisted_plan_ledger["completed_gates"] == ["clarification_complete", "plan_executable"]
-    assert persisted_plan_ledger["tasks"][0]["evidence_refs"] == ["run-state.json#gate:plan_executable"]
-
-    result = run_route(task_dir=task_dir, policy=policy, route_name="medium")
-
-    assert result["current_stage"] == "finalize"
-    decision_log = json.loads((task_dir / "decision-log.json").read_text(encoding="utf-8"))
-    decisions = [entry["decision"] for entry in decision_log["entries"]]
-    assert "step back: quality-review -> execute" in decisions
-    assert decisions[-4:] == [
-        "route resumed: medium from execute",
-        "stage entered: execute",
-        "stage entered: quality-review",
-        "stage entered: finalize",
-    ]
-
-
-def test_step_back_large_route_preserves_spec_evidence_and_clears_quality_flag(tmp_path: Path) -> None:
-    policy = load_runtime_policy(ROOT)
-    task_dir = _make_task_dir(tmp_path)
-    _write_medium_plan_artifacts(task_dir, route_name="large_high_risk")
-    _write_json(
-        task_dir / "run-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "current_stage": "long-run",
-            "status": "in_progress",
-            "completed_gates": ["clarification_complete"],
-            "failed_gates": [],
-            "retries": {},
-            "current_task_id": "task-1",
-            "spec_review_approved": True,
-            "quality_review_approved": True,
-        },
-    )
-    _write_json(
-        task_dir / "plan-ledger.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "route": "large_high_risk",
-            "completed_stages": ["clarify", "plan", "execute", "spec-review", "quality-review", "finalize"],
-            "completed_gates": [
-                "clarification_complete",
-                "plan_executable",
-                "execution_evidenced",
-                "spec_review_passed",
-                "quality_review_passed",
-                "ready_to_finalize",
-            ],
-            "retries": {},
-            "current_task_id": "task-1",
-            "last_review_verdict": "approved",
-            "tasks": [
-                {
-                    "id": "task-1",
-                    "title": "update workflow docs",
-                    "depends_on": [],
-                    "files": ["docs/workflow.md"],
-                    "parallel_safe": False,
-                    "status": "in_progress",
-                    "required_gates": ["machine", "validator"],
-                    "evidence_refs": [
-                        "run-state.json#gate:plan_executable",
-                        "run-state.json#gate:execution_evidenced",
-                        "review-report-spec.json#verdict:approved",
-                        "review-report-quality.json#verdict:approved",
-                        "run-state.json#gate:quality_review_passed",
-                        "run-state.json#gate:ready_to_finalize",
-                        "eval-record.json#verdict:approved",
-                    ],
-                    "attempt_count": 1,
-                }
-            ],
-        },
-    )
-
-    state = step_back(task_dir=task_dir, policy=policy, route_name="large_high_risk", current_stage="long-run")
-
-    assert state["current_stage"] == "finalize"
-    assert state["spec_review_approved"] is True
-    assert state["quality_review_approved"] is False
-    persisted_plan_ledger = json.loads((task_dir / "plan-ledger.json").read_text(encoding="utf-8"))
-    assert "review-report-spec.json#verdict:approved" in persisted_plan_ledger["tasks"][0]["evidence_refs"]
-    assert "review-report-quality.json#verdict:approved" in persisted_plan_ledger["tasks"][0]["evidence_refs"]
-    assert "eval-record.json#verdict:approved" not in persisted_plan_ledger["tasks"][0]["evidence_refs"]
 
 
 def test_escalate_route_switches_to_large_high_risk(tmp_path: Path) -> None:
